chore: refactor utils into more specific files
